refactored error handling. all being handled in ClientMain
