+++ song-client/src/main/java/org/icgc/dcc/song/client/cli/ClientMain.java
@@ -28,22 +28,34 @@ import org.icgc.dcc.song.client.command.SearchCommand;
 import org.icgc.dcc.song.client.command.StatusCommand;
 import org.icgc.dcc.song.client.command.UploadCommand;
 import org.icgc.dcc.song.client.config.Config;
+import org.icgc.dcc.song.client.register.ErrorStatusHeader;
 import org.icgc.dcc.song.client.register.Registry;
+import org.icgc.dcc.song.core.exceptions.ServerException;
+import org.icgc.dcc.song.core.exceptions.SongError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestClientException;
+
+import java.net.HttpRetryException;
+
+import static org.icgc.dcc.song.core.exceptions.ServerErrors.UNAUTHORIZED_TOKEN;
+import static org.icgc.dcc.song.core.exceptions.ServerErrors.UNKNOWN_ERROR;
+import static org.icgc.dcc.song.core.exceptions.SongError.createSongError;
 
 @Component
 @Slf4j
 public class ClientMain implements CommandLineRunner {
 
   private CommandParser dispatcher;
+  private ErrorStatusHeader errorStatusHeader;
+  private Registry registry;
+  private Config config;
 
   @Autowired
   ClientMain(Config config, Registry registry) {
     val programName = config.getProgramName();
     val options = new Options();
-
     val builder = new CommandParserBuilder(programName, options);
     builder.register("config", new ConfigCommand(config));
     builder.register("upload", new UploadCommand(registry));
@@ -53,13 +65,47 @@ public class ClientMain implements CommandLineRunner {
     builder.register("manifest", new ManifestCommand(registry, config));
     builder.register("publish", new PublishCommand(registry, config));
     this.dispatcher = builder.build();
+    this.errorStatusHeader = new ErrorStatusHeader(config);
+    this.registry = registry;
+    this.config = config;
   }
 
   @Override
   public void run(String... args) {
     val command = dispatcher.parse(args);
-    command.run();
-    command.report();
+    try{
+      command.run();
+    } catch(RestClientException e){
+      val isAlive = registry.isServerAlive();
+      SongError songError;
+      if(isAlive){
+        val cause = e.getCause();
+        if (cause instanceof  HttpRetryException){
+          val httpRetryException = (HttpRetryException)cause;
+          if (httpRetryException.responseCode() == UNAUTHORIZED_TOKEN.getHttpStatus().value()){
+            songError = createSongError(UNAUTHORIZED_TOKEN,"Invalid token");
+          } else {
+            songError = createSongError(UNKNOWN_ERROR,
+                "Unknown error with ResponseCode [%s] -- Reason: %s, Message: %s",
+                httpRetryException.responseCode(),
+                httpRetryException.getReason(),
+                httpRetryException.getMessage());
+          }
+        } else {
+          songError = createSongError(UNKNOWN_ERROR,
+              "Unknown error: %s", e.getMessage());
+        }
+        command.err(errorStatusHeader.getSongServerErrorOutput(songError));
+      } else {
+        command.err(errorStatusHeader.createMessage("The SONG server may not be running on '%s'", config.getServerUrl()));
+      }
+    } catch (ServerException ex){
+      val songError = ex.getSongError();
+      command.err(errorStatusHeader.getSongServerErrorOutput(songError));
+    } finally{
+      command.report();
+    }
+
   }
 
 }
