+++ song-client/src/main/java/org/icgc/dcc/song/client/register/Registry.java
@@ -24,51 +24,26 @@ import lombok.SneakyThrows;
 import lombok.val;
 import org.icgc.dcc.song.client.cli.Status;
 import org.icgc.dcc.song.client.config.Config;
-import org.icgc.dcc.song.core.exceptions.SongError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.ResourceAccessException;
-
-import java.net.HttpRetryException;
-import java.util.function.Supplier;
 
 import static java.lang.Boolean.parseBoolean;
-import static org.icgc.dcc.song.core.exceptions.ServerErrors.UNAUTHORIZED_TOKEN;
-import static org.icgc.dcc.song.core.exceptions.ServerErrors.UNKNOWN_ERROR;
-import static org.icgc.dcc.song.core.exceptions.SongError.createSongError;
 
 @Component
 public class Registry {
 
-  private static final boolean DEFAULT_DEBUG_ENABLE = false;
-  
   @Setter
   private RestClient restClient;
   private ObjectMapper mapper;
   private Endpoint endpoint;
   private String accessToken;
-  private ErrorStatusHeader errorStatusHeader;
-  private Config config;
 
   @Autowired
-  public Registry(Config config, RestClient restClient, ErrorStatusHeader errorStatusHeader) {
+  public Registry(Config config, RestClient restClient) {
     this.mapper = new ObjectMapper();
     this.restClient = restClient;
     this.endpoint = new Endpoint(config.getServerUrl());
     this.accessToken = config.getAccessToken();
-    this.errorStatusHeader = errorStatusHeader;
-    this.config = config;
-  }
-
-
-  @SneakyThrows
-  String getAnalysisType(String json) {
-    val node = mapper.readTree(json);
-
-    if (node.has("analysisType")) {
-      return node.get("analysisType").asText();
-    }
-    throw new Error("No analysis type specified in JSON document" + node.asText());
   }
 
   @SneakyThrows
@@ -76,7 +51,6 @@ public class Registry {
     val node = mapper.readTree(json);
     return node.get("study").asText();
   }
-
 
   /**
    * Register an analysis with the song server.
@@ -86,7 +60,7 @@ public class Registry {
    */
   public Status upload(String json, boolean isAsyncValidation) {
     val url = endpoint.upload(getStudyId(json), isAsyncValidation);
-    return tryPostAuth(accessToken, url, json);
+    return restClient.postAuth(accessToken, url, json);
   }
 
   /***
@@ -97,17 +71,17 @@ public class Registry {
    */
   public Status getUploadStatus(String studyId, String uploadId) {
     val url = endpoint.status(studyId, uploadId);
-    return tryGet(accessToken, url);
+    return restClient.get(accessToken, url);
   }
 
   public Status save(String studyId, String uploadId) {
     val url = endpoint.saveById(studyId, uploadId);
-    return tryPostAuth(accessToken, url);
+    return restClient.postAuth(accessToken, url);
   }
 
   public Status getAnalysisFiles(String studyId, String analysisId) {
     val url = endpoint.getAnalysisFiles(studyId, analysisId);
-    return tryGet(accessToken, url);
+    return restClient.get(accessToken, url);
   }
 
   /**
@@ -129,7 +103,7 @@ public class Registry {
    */
   public Status publish(String studyId, String analysisId ){
     val url = endpoint.publish(studyId, analysisId);
-    return tryPutAuth(accessToken, url);
+    return restClient.putAuth(accessToken, url);
   }
 
   public Status search(String studyId,
@@ -138,61 +112,7 @@ public class Registry {
       String donorId,
       String fileId){
     val url = endpoint.searchGet(studyId,sampleId,specimenId,donorId,fileId);
-    return tryGet(accessToken, url);
-  }
-
-  /**
-   * Wraps the restClient calls, so that unauthorized.token errors are handled properly. In addition, if an error occurs, the availability of the song server is checked, to ensure its running.
-   * @param supplier Request method that returns a Status object
-   * @return the status
-   */
-  private Status tryClient(Supplier<Status> supplier){
-    val status = new Status();
-    try {
-      return supplier.get();
-    } catch(ResourceAccessException e){
-      val isAlive = isServerAlive();
-      SongError songError;
-      if(isAlive){
-        val cause = e.getCause();
-        if (cause instanceof HttpRetryException){
-          val httpRetryException = (HttpRetryException)cause;
-          if (httpRetryException.responseCode() == UNAUTHORIZED_TOKEN.getHttpStatus().value()){
-            songError = createSongError(UNAUTHORIZED_TOKEN,"Invalid token");
-          } else {
-            songError = createSongError(UNKNOWN_ERROR,
-                "Unknown error with ResponseCode [%s] -- Reason: %s, Message: %s",
-                httpRetryException.responseCode(),
-                httpRetryException.getReason(),
-                httpRetryException.getMessage());
-          }
-        } else {
-          songError = createSongError(UNKNOWN_ERROR,
-              "Unknown error: %s", e.getMessage());
-        }
-        status.err(errorStatusHeader.getSongServerErrorOutput(songError));
-      } else {
-        status.err(errorStatusHeader.createMessage("The SONG server may not be running on '%s'. Ensure "
-            + "the GET response of the '%s' endpoint returns true", config.getServerUrl(), endpoint.isAlive()));
-      }
-      return status;
-    }
-  }
-
-  private Status tryGet(String accessToken, String url){
-    return tryClient(() -> restClient.get(accessToken, url));
-  }
-
-  private Status tryPutAuth(String accessToken, String url){
-    return tryClient(() -> restClient.putAuth(accessToken, url));
-  }
-
-  private Status tryPostAuth(String accessToken, String url){
-    return tryClient(() -> restClient.postAuth(accessToken, url));
-  }
-
-  private Status tryPostAuth(String accessToken, String url, String json){
-    return tryClient(() -> restClient.postAuth(accessToken, url, json));
+    return restClient.get(accessToken, url);
   }
 
 }
